requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
Flask-Limiter==3.5.0
Flask-Mail==0.9.1
pymongo==4.5.0
bcrypt==4.0.1
python-dotenv==1.0.0
Pillow==10.0.1
cloudinary==1.36.0
marshmallow==3.20.1
marshmallow-mongoengine==0.29.0
email-validator==2.0.0
gunicorn==21.2.0
python-multipart==0.0.6

# MongoDB and Testing Dependencies
motor==3.3.2
pytest-mongodb==2.0.0
pytest-asyncio==0.21.1
mongomock==4.1.2



